pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "autochangelog"
version = "0.1.1"
description = "Automated changelog generation using LLMs and commit history analysis."
readme = "README.md"
requires-python = ">3.10"
dependencies = [
  "requests==2.32.5",
  "langchain==1.2.3",
  "langchain-litellm",
  "rich==14.2.0",
  "python-dotenv"
]
license = {text = "MIT"}
authors = [{name = "Milansuman", email = "milansuman2016@gmail.com"}]

[tool.setuptools.packages.find]
where = ["."]
include = ["autochangelog*"]
exclude = ["autochangelog.tests*"]

[project.scripts]
autochangelog = "autochangelog.agent:main"